refactor: split events.rs into per-domain submodules
Split the monolithic 1956-line events.rs into domain-specific submodules:
- events/project.rs: project, claim, bookmark, follow, endorsement events
- events/review.rs: review and review moderation events
- events/verification.rs: verification, renewal, dispute events
- events/fee.rs: fee-related events
- events/admin.rs: admin, timelock, collection, pause events
- events/mod.rs: re-exports all public items for backward compatibility

All existing imports via 'use crate::events::*' continue to work unchanged.
