+++ dongle-smartcontract/src/events/mod.rs
@@ -1,0 +1,11 @@
+pub mod admin;
+pub mod fee;
+pub mod project;
+pub mod review;
+pub mod verification;
+
+pub use admin::*;
+pub use fee::*;
+pub use project::*;
+pub use review::*;
+pub use verification::*;
